+++ node/clients/storefrontPermissions.ts
@@ -314,7 +314,9 @@ export default class StorefrontPermissions extends AppGraphQLClient {
   }
 
   private getTokenToHeader = () => {
-    const adminToken = this.context.adminUserAuthToken ?? this.context.authToken
+    // provide authToken (app token) as an admin token as this is a call
+    // between b2b suite apps and no further token validation is needed
+    const adminToken = this.context.authToken
     const userToken = this.context.storeUserAuthToken ?? null
     const { sessionToken, account } = this.context