+++ node/resolvers/directives/checkAdminAccess.ts
@@ -4,6 +4,11 @@ import type { GraphQLField } from 'graphql'
 import { defaultFieldResolver } from 'graphql'
 
 import sendAuthMetric, { AuthMetric } from '../../utils/metrics/auth'
+import {
+  validateAdminToken,
+  validateAdminTokenOnHeader,
+  validateApiToken,
+} from './helper'
 
 export class CheckAdminAccess extends SchemaDirectiveVisitor {
   public visitFieldDefinition(field: GraphQLField<any, any>) {
@@ -16,55 +21,87 @@ export class CheckAdminAccess extends SchemaDirectiveVisitor {
       info: any
     ) => {
       const {
-        vtex: { adminUserAuthToken, logger },
-        clients: { identity },
+        vtex: { adminUserAuthToken, storeUserAuthToken, logger },
       } = context
 
-      // send metric to collect data about caller and tokens used in the request
-      const operation = field.astNode?.name?.value ?? context?.request?.url
-      const metric = new AuthMetric(
+      const { hasAdminToken, hasValidAdminToken, hasCurrentValidAdminToken } =
+        await validateAdminToken(context, adminUserAuthToken as string)
+
+      const {
+        hasAdminTokenOnHeader,
+        hasValidAdminTokenOnHeader,
+        hasCurrentValidAdminTokenOnHeader,
+      } = await validateAdminTokenOnHeader(context)
+
+      const { hasApiToken, hasValidApiToken, hasCurrentValidApiToken } =
+        await validateApiToken(context)
+
+      const hasStoreToken = !!storeUserAuthToken // we don't need to validate store token
+
+      // now we emit a metric with all the collected data before we proceed
+      const operation = field.astNode?.name?.value ?? context.request.url
+      const userAgent = context?.request?.headers['user-agent'] as string
+      const caller = context?.request?.headers['x-vtex-caller'] as string
+      const forwardedHost = context?.request?.headers[
+        'x-forwarded-host'
+      ] as string
+
+      const auditMetric = new AuthMetric(
         context?.vtex?.account,
         {
-          caller: context?.request?.headers['x-vtex-caller'] as string,
-          userAgent: context?.request?.headers['user-agent'] as string,
-          forwardedHost: context?.request?.headers[
-            'x-forwarded-host'
-          ] as string,
-          hasAdminToken: !!adminUserAuthToken,
-          hasApiToken: !!context?.request?.headers['vtex-api-apptoken'],
-          hasStoreToken: false,
           operation,
+          forwardedHost,
+          caller,
+          userAgent,
+          hasAdminToken,
+          hasValidAdminToken,
+          hasApiToken,
+          hasValidApiToken,
+          hasStoreToken,
+          hasAdminTokenOnHeader,
+          hasValidAdminTokenOnHeader,
         },
-        'CheckAdminAccess'
+        'CheckAdminAccessAudit'
       )
 
-      sendAuthMetric(context, logger, metric)
+      sendAuthMetric(context, logger, auditMetric)
 
-      let token =
-        adminUserAuthToken ?? (context?.headers.vtexidclientautcookie as string)
-
-      const apiToken = context?.headers['vtex-api-apptoken'] as string
-      const appKey = context?.headers['vtex-api-appkey'] as string
-
-      if (apiToken?.length && appKey?.length) {
-        token = (
-          await identity.getToken({ appkey: appKey, apptoken: apiToken })
-        ).token
-        context.cookies.set('VtexIdclientAutCookie', token)
-        context.vtex.adminUserAuthToken = token
-      }
-
-      if (!token) {
+      if (!hasAdminToken && !hasApiToken && !hasAdminTokenOnHeader) {
+        logger.warn({
+          message: 'CheckAdminAccess: No token provided',
+          userAgent,
+          caller,
+          forwardedHost,
+          operation,
+          hasAdminToken,
+          hasValidAdminToken,
+          hasApiToken,
+          hasValidApiToken,
+          hasStoreToken,
+          hasAdminTokenOnHeader,
+          hasValidAdminTokenOnHeader,
+        })
         throw new AuthenticationError('No token was provided')
       }
 
-      try {
-        await identity.validateToken({ token })
-      } catch (err) {
+      if (
+        !hasCurrentValidAdminToken &&
+        !hasCurrentValidApiToken &&
+        !hasCurrentValidAdminTokenOnHeader
+      ) {
         logger.warn({
-          error: err,
           message: 'CheckAdminAccess: Invalid token',
-          token,
+          userAgent,
+          caller,
+          forwardedHost,
+          operation,
+          hasAdminToken,
+          hasValidAdminToken,
+          hasApiToken,
+          hasValidApiToken,
+          hasStoreToken,
+          hasAdminTokenOnHeader,
+          hasValidAdminTokenOnHeader,
         })
         throw new ForbiddenError('Unauthorized Access')
       }