+++ CHANGELOG.md
@@ -8,17 +8,103 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Fixed
-
 - Fix deleteCostCenter mutation
+
+## [0.57.0] - 2024-09-12
+
+### Added
+- Add new admin token validation directive to some APIs
+
+## [0.56.2] - 2024-09-10
+
+### Fixed
+
+- Remove error log when file not found for marketing's tags.
+
+## [0.56.1] - 2024-09-10
+
+### Added
+- Add logging for store token validation
+
+## [0.56.0] - 2024-09-09
+
+### Added
+
+Adds a new `getAccount` query to retrieve account information. It includes fields such as `id`, `name`, `accountName`, `tradingName`, and `isActive`. This query is implemented in both the GraphQL schema and the LMClient class.
+
+## [0.55.0] - 2024-08-22
+
+### Added
+- Add paymentTerms field to cost center input on mutations
+
+## [0.54.0] - 2024-08-12
+
+### Added
+
+- Add new `getSellersPaginated` query to allow pagination on sellers query
+
+## [0.53.0] - 2024-08-06
+
+### Fixed
+
+- Add paymentTerms field on create cost center mutation
+
+## [0.52.0] - 2024-08-05
+
+### Added
+
+- New admin user token validation directive
+
+## [0.51.2] - 2024-07-30
+
+### Fixed
+
+- Provide app token on calls to storefront-permissions app
+
+## [0.51.1] - 2024-07-22
+
+### Added
+
+- Audit metrics for some graphql APIs
+- Improve access directives
+
+## [0.51.0] - 2024-06-04
+
+### Fixed
+
+- Removed 0.50.0 version changes which contained a bug
+
+## [0.50.0] - 2024-06-03
+
+### Added
+
+- `Trade Name` to be searchable (Admin)
+- `Trade Name` to the Organization listing page (Admin)
+- `Trade Name` to the Organization Requests listing page (Admin)
+- Event listener to update schemas once new app version is released
+
+## [0.49.6] - 2024-06-03
+
+### Fixed
+
+- Fix check access directive by allowing appkey tokens for authentication
+
+## [0.49.5] - 2024-05-29
+
+### Changed
+
+- Improve logging and metrics for checkUserAccess and checkAdminAccess directives
 
 ## [0.49.4] - 2024-05-07
 
 ### Added
+
 - Add metric to check access directives
 
 ## [0.49.3] - 2024-04-24
 
 ### Fixed
+
 - Provide correct tokens to clients
 
 ## [0.49.2] - 2024-04-18
@@ -60,11 +146,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.48.2] - 2024-02-21
 
 ### Fixed
+
 - Add analytics client to properly send metrics to redshift
 
 ## [0.48.1] - 2024-02-07
 
 ### Fixed
+
 - Add retries to `sendMetric` to avoid connection errors
 
 ## [0.48.0] - 2024-01-19
@@ -72,6 +160,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Allow `paymentTerms`, `collections` and `sellers` to be provided by name only on `createOrganizationAndCostCenterWithAdminUser`
+
 ## [0.47.1] - 2023-12-18
 
 ### Fixed
@@ -158,7 +247,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.37.0] - 2023-09-19
 
-
 ### Added
 
 - Added event trigger on delete a user
@@ -174,7 +262,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added metrics for impersonate user
-- 
+-
+
 ## [0.35.3] - 2023-08-09
 
 ### Fixed
@@ -192,7 +281,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fix auth problem when fetching orders history
-
 
 ## [0.35.0] - 2023-07-24